j606f: remove usage of custom audio dspg usecase and device
Those previously were enabled once the screen is on and an application
requests a voice recognition source. Remove and add them to the
SND_DEVICE_IN_VOICE_REC_MIC device which is basically selected by the
same condition but without the screen off criteria. Also use a custom
backend in order to select the correct modified path.